Cleanup, review suggestions, namespaces
