+++ packages/expo-av/ios/EXAV/AudioSampleCallback/EXAV+AudioSampleCallback.h
@@ -3,12 +3,13 @@
 #pragma once
 
 #import <Foundation/Foundation.h>
+
 #import <EXAV/EXAV.h>
 #import <EXAV/EXAVPlayerData.h>
 
 @interface EXAV (AudioSampleCallback)
 
-- (void) installJSIBindingsForRuntime:(void *)jsRuntimePtr
-                  withSoundDictionary:(NSMutableDictionary<NSNumber*, EXAVPlayerData*>*)soundDictionary;
+- (void)installJSIBindingsForRuntime:(void *)jsRuntimePtr
+                 withSoundDictionary:(NSMutableDictionary<NSNumber*, EXAVPlayerData*>*)soundDictionary;
 
 @end
